+++ unstructured/ingest/interfaces.py
@@ -3,6 +3,7 @@ through Unstructured."""
 
 from abc import ABC, abstractmethod
 
+from unstructured.ingest.logger import logger
 from unstructured.partition.auto import partition
 from unstructured.staging.base import convert_to_dict
 
@@ -46,7 +47,6 @@ class BaseConnectorConfig(ABC):
     ***REMOVED*** where to write structured data outputs
     output_dir: str
     re_download: bool = False
-    verbose: bool = False
 
 
 class BaseIngestDoc(ABC):
@@ -86,7 +86,7 @@ class BaseIngestDoc(ABC):
         pass
 
     def process_file(self):
-        print(f"Processing {self.filename}")
+        logger.info(f"Processing {self.filename}")
 
         elements = partition(filename=str(self.filename))
         isd_elems = convert_to_dict(elements)