+++ unstructured/ingest/main.py
@@ -1,5 +1,6 @@
 ***REMOVED***!/usr/bin/env python3
 import hashlib
+import logging
 import multiprocessing as mp
 import sys
 from pathlib import Path
@@ -18,18 +19,28 @@ from unstructured.ingest.connector.wikipedia import (
     WikipediaConnector,
 )
 from unstructured.ingest.doc_processor.generalized import initialize, process_document
+from unstructured.ingest.logger import ingest_log_streaming_init, logger
 
 
 class MainProcess:
-    def __init__(self, doc_connector, doc_processor_fn, num_processes, reprocess):
+    def __init__(
+        self,
+        doc_connector,
+        doc_processor_fn,
+        num_processes,
+        reprocess,
+        verbose,
+    ):
         ***REMOVED*** initialize the reader and writer
         self.doc_connector = doc_connector
         self.doc_processor_fn = doc_processor_fn
         self.num_processes = num_processes
         self.reprocess = reprocess
+        self.verbose = verbose
 
     def initialize(self):
         """Slower initialization things: check connections, load things into memory, etc."""
+        ingest_log_streaming_init(logging.DEBUG if self.verbose else logging.INFO)
         self.doc_connector.initialize()
         initialize()
 
@@ -41,12 +52,12 @@ class MainProcess:
         docs = [doc for doc in docs if not doc.has_output()]
         num_docs_to_process = len(docs)
         if num_docs_to_process == 0:
-            print(
+            logger.info(
                 "All docs have structured outputs, nothing to do. Use --reprocess to process all.",
             )
             return None
         elif num_docs_to_process != num_docs_all:
-            print(
+            logger.info(
                 f"Skipping processing for {num_docs_all - num_docs_to_process} docs out of "
                 f"{num_docs_all} since their structured outputs already exist, use --reprocess to "
                 "reprocess those in addition to the unprocessed ones.",
@@ -69,8 +80,12 @@ class MainProcess:
         ***REMOVED*** block to remain in single process
         ***REMOVED*** self.doc_processor_fn(docs[0])
 
-        with mp.Pool(processes=self.num_processes) as pool:
-            results = pool.map(self.doc_processor_fn, docs)  ***REMOVED*** noqa: F841
+        with mp.Pool(
+            processes=self.num_processes,
+            initializer=ingest_log_streaming_init,
+            initargs=(logging.DEBUG if self.verbose else logging.INFO,),
+        ) as pool:
+            pool.map(self.doc_processor_fn, docs)  ***REMOVED*** noqa: F841
 
         self.cleanup()
 
@@ -239,7 +254,9 @@ def main(
     verbose,
 ):
     if not preserve_downloads and download_dir:
-        print("Warning: not preserving downloaded files but --download_dir is specified")
+        logger.warning(
+            "Not preserving downloaded files but --download_dir is specified",
+        )
     if not download_dir:
         cache_path = Path.home() / ".cache" / "unstructured" / "ingest"
         if not cache_path.exists():
@@ -266,8 +283,8 @@ def main(
             raise ValueError("No connector-specific option was specified!")
         download_dir = cache_path / hashed_dir_name.hexdigest()[:10]
         if preserve_downloads:
-            print(
-                f"Warning: preserving downloaded files but --download-dir is not specified,"
+            logger.warning(
+                f"Preserving downloaded files but --download-dir is not specified,"
                 f" using {download_dir}",
             )
     if s3_url:
@@ -280,7 +297,6 @@ def main(
                 anonymous=s3_anonymous,
                 re_download=re_download,
                 preserve_downloads=preserve_downloads,
-                verbose=verbose,
             ),
         )
     elif github_url:
@@ -295,7 +311,6 @@ def main(
                 preserve_downloads=preserve_downloads,
                 output_dir=structured_output_dir,
                 re_download=re_download,
-                verbose=verbose,
             ),
         )
     elif subreddit_name:
@@ -312,7 +327,6 @@ def main(
                 preserve_downloads=preserve_downloads,
                 output_dir=structured_output_dir,
                 re_download=re_download,
-                verbose=verbose,
             ),
         )
     elif wikipedia_page_title:
@@ -325,7 +339,6 @@ def main(
                 preserve_downloads=preserve_downloads,
                 output_dir=structured_output_dir,
                 re_download=re_download,
-                verbose=verbose,
             ),
         )
     elif drive_id:
@@ -340,14 +353,13 @@ def main(
                 preserve_downloads=preserve_downloads,
                 output_dir=structured_output_dir,
                 re_download=re_download,
-                verbose=verbose,
             ),
         )
     ***REMOVED*** Check for other connector-specific options here and define the doc_connector object
     ***REMOVED*** e.g. "elif azure_container:  ..."
 
     else:
-        print("No connector-specific option was specified!")
+        logger.error("No connector-specific option was specified!")
         sys.exit(1)
 
     MainProcess(
@@ -355,6 +367,7 @@ def main(
         doc_processor_fn=process_document,
         num_processes=num_processes,
         reprocess=reprocess,
+        verbose=verbose,
     ).run()